Upgrade existing strategy test
